requirements.txt
requests
rich
python-slugify
Jinja2
